% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustperm.R
\name{aov_by_bin}
\alias{aov_by_bin}
\title{Bin-by-bin analysis of variance}
\usage{
aov_by_bin(.data, bin, formula)
}
\arguments{
\item{.data}{Data frame.}

\item{bin}{Unquoted name of variable identifying bins.}

\item{formula}{Model formula, passed to \code{\link{aov}}.}
}
\value{
A data frame containing signed F statistics and p values for each time bin.
TODO examples
}
\description{
Run \code{\link{aov}} on each of a series of time bins.
}
