% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustperm.R
\name{cluster_nhds}
\alias{cluster_nhds}
\title{Generate null-hypothesis distributions for cluster mass statistics}
\usage{
cluster_nhds(n, .data, bin, formula, fn, ...)
}
\arguments{
\item{n}{Number of Monte Carlo runs}

\item{.data}{Data frame.}

\item{bin}{Unquoted name of variable identifying bins.}

\item{formula}{Model formula, passed to \code{\link{aov_by_bin}}.}

\item{fn}{Unquoted name of function to use for exchanging labels.}

\item{...}{Arguments to be passed to \code{fn}.}
}
\value{
A data frame with null-hypothesis distributions for each effect in the column \code{nhd}.
}
\description{
Generate null-hypothesis distributions for cluster mass statistics
}
\details{
Generates null-hypothesis distributions by the following algorithm: For each of the \code{n} Monte Carlo runs, (1) randomly permute labels in the dataset according to the relabeling function \code{fn}; (2) run \code{\link{aov_by_bin}} on the resulting data, the result of which is passed to (3) \code{\link{detect_clusters_by_effect}}, and (4) store the maximum cluster mass statistic for each effect on each run.
}
