% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustperm.R
\name{detect_clusters}
\alias{detect_clusters}
\title{Detect clusters and calculate mass statistics}
\usage{
detect_clusters(.data, bin, stat, p, alpha = 0.05)
}
\arguments{
\item{.data}{Data frame.}

\item{bin}{Unquoted name of variable identifying bins.}

\item{stat}{Unquoted name of variable with signed statistics.}

\item{p}{Unquoted name of variable with p values.}

\item{alpha}{Alpha level for each test.}
}
\value{
A data frame listing onset (\code{b0}) and offset (\code{b0}) of each cluster, along with the sign of the effect and cluster mass statistic (\code{cms}).
}
\description{
Detect clusters and calculate mass statistics
}
\seealso{
\code{\link{detect_clusters_by_effect}} for more than one effect.
}
