% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustperm.R
\name{detect_clusters_by_effect}
\alias{detect_clusters_by_effect}
\title{Detect clusters and calculate mass statistics for multiple effects}
\usage{
detect_clusters_by_effect(.data, effect, bin, stat, p, alpha = 0.05)
}
\arguments{
\item{.data}{Data frame.}

\item{effect}{Unquoted name of variable identifying each effect.}

\item{bin}{Unquoted name of variable identifying bins.}

\item{stat}{Unquoted name of variable with signed statistics.}

\item{p}{Unquoted name of variable with p values.}

\item{alpha}{Alpha level for each test.}
}
\value{
Data frame with effects, cluster onsets (\code{b0}) and offsets (\code{b1}) and cluster mass statistics (\code{cms}).
}
\description{
Detect clusters and calculate mass statistics for multiple effects
}
\details{
Performs \code{\link{detect_clusters}} for each individual effect in the data frame.
}
