% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustperm.R
\name{.eff_sign}
\alias{.eff_sign}
\title{Get the sign of a difference}
\usage{
.eff_sign(x)
}
\arguments{
\item{x}{A vector of length 2, a 2x2 matrix, or a 2x2x2 matrix.}
}
\value{
An integer (1 or -1), indicating effect direction.
}
\description{
Get the sign of a difference
}
\details{
A convenience function that takes the output of \code{\link{model.tables}} applied to an \code{\link{aov}} object. Not exported.
}
