% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustperm.R
\name{.tidy_anova}
\alias{.tidy_anova}
\title{Run an analysis of variance}
\usage{
.tidy_anova(x, formula)
}
\arguments{
\item{x}{Data frame (a single bin).}

\item{formula}{Model formula (passed to \code{\link{aov}}). For within-subjects data, and \code{Error} term should be simple (e.g., \code{Error(id)}.}
}
\value{
A data frame with effect, signed F statistic (indicating direction of the effect), and p value.
}
\description{
Run an analysis of variance
}
