% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustperm.R
\name{pvalue}
\alias{pvalue}
\title{Calculate permutation p-value from null-hypothesis distribution}
\usage{
pvalue(cms, nhd)
}
\arguments{
\item{cms}{Cluster mass statistic for the original data.}

\item{nhd}{A vector comprising the null-hypothesis distribution, excluding the original statistic.}
}
\value{
A scalar representing the p-value.
}
\description{
Calculate permutation p-value from null-hypothesis distribution
}
\details{
Calculates the p-value as the number of values in the vector \code{c(cms, nhd)} that are greater than or equal to the original value, divided by the length of \code{nhd} plus one (because the original statistic is included in the NHD).
}
