% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustperm.R
\name{pvalues}
\alias{pvalues}
\title{Calculate permutation p-values from null-hypothesis distributions}
\usage{
pvalues(orig, nhds, key = "effect")
}
\arguments{
\item{orig}{Data frame containing clusters for multiple effects detected on original data.}

\item{nhds}{Data frame with nhds for each effect.}

\item{key}{Key on which to join the data frames (i.e., quoted name of the variable identifying effects).}
}
\value{
Original data frame with extra column containing p-values for each effect and cluster.
}
\description{
Calculate permutation p-values from null-hypothesis distributions
}
